refactor: split store_test.go (2332→20) and migrations_test.go into 9 per-version test files
