+++ internal/store/004_dedup_test.go
@@ -1,0 +1,58 @@
+package store
+
+import (
+	"database/sql"
+	"path/filepath"
+	"testing"
+
+	"github.com/andrey-vk/wdbgp/internal/config"
+
+	_ "modernc.org/sqlite"
+)
+
+func TestMigrationDeduplicatesFeedsWithoutLosingCatalog(t *testing.T) {
+	path := filepath.Join(t.TempDir(), "duplicate-feeds.sqlite3")
+	db, err := sql.Open("sqlite", path)
+	if err != nil {
+		t.Fatal(err)
+	}
+	_, err = db.Exec(`
+CREATE TABLE feeds (
+ id INTEGER PRIMARY KEY, name TEXT NOT NULL UNIQUE, url TEXT NOT NULL,
+ enabled INTEGER NOT NULL DEFAULT 1, last_success TEXT, last_error TEXT
+);
+CREATE TABLE catalog_entries (
+ feed_id INTEGER NOT NULL REFERENCES feeds(id) ON DELETE CASCADE,
+ category TEXT NOT NULL, service TEXT NOT NULL, cidr TEXT NOT NULL,
+ PRIMARY KEY (feed_id, category, service, cidr)
+);
+INSERT INTO feeds(id, name, url) VALUES
+ (1, 'old-name', 'https://example.test/feed'),
+ (2, 'new-name', 'https://example.test/feed');
+INSERT INTO catalog_entries VALUES
+ (1, 'one', 'a', '192.0.2.0/24'),
+ (2, 'two', 'b', '198.51.100.0/24');
+`)
+	if err != nil {
+		t.Fatal(err)
+	}
+	db.Close()
+
+	s, err := Open(path, config.Config{})
+	if err != nil {
+		t.Fatal(err)
+	}
+	defer s.Close()
+	var feedCount, entryCount int
+	if err := s.DB.QueryRow("SELECT COUNT(*) FROM feeds WHERE url = 'https://example.test/feed'").
+		Scan(&feedCount); err != nil {
+		t.Fatal(err)
+	}
+	if err := s.DB.QueryRow("SELECT COUNT(*) FROM catalog_entries WHERE feed_id = 1").
+		Scan(&entryCount); err != nil {
+		t.Fatal(err)
+	}
+	if feedCount != 1 || entryCount != 2 {
+		t.Fatalf("feeds=%d entries=%d, want 1 and 2", feedCount, entryCount)
+	}
+}
