+++ internal/store/005_filters_test.go
@@ -1,0 +1,96 @@
+package store
+
+import (
+	"context"
+	"database/sql"
+	"path/filepath"
+	"testing"
+
+	"github.com/andrey-vk/wdbgp/internal/config"
+
+	_ "modernc.org/sqlite"
+)
+
+func TestMigrationPreservesLegacyFilterOverrideMode(t *testing.T) {
+	path := filepath.Join(t.TempDir(), "legacy-filter-mode.sqlite3")
+	db, err := sql.Open("sqlite", path)
+	if err != nil {
+		t.Fatal(err)
+	}
+	_, err = db.Exec(`
+CREATE TABLE schema_migrations (
+    version INTEGER PRIMARY KEY, name TEXT NOT NULL, applied_at TEXT NOT NULL
+);
+INSERT INTO schema_migrations(version, name, applied_at) VALUES
+    (1, 'initial schema', 'now'),
+    (2, 'add OpenCCK IPv6 feeds', 'now'),
+    (3, 'add lookup indexes', 'now'),
+    (4, 'deduplicate feeds by URL', 'now'),
+    (5, 'add route filters', 'now');
+CREATE TABLE users (
+    id INTEGER PRIMARY KEY,
+    name TEXT NOT NULL UNIQUE,
+    peer_ip TEXT NOT NULL UNIQUE,
+    peer_asn INTEGER NOT NULL,
+    next_hop TEXT,
+    bgp_password TEXT,
+    selection_locked INTEGER NOT NULL DEFAULT 0,
+    enabled INTEGER NOT NULL DEFAULT 1,
+    filter_override_enabled INTEGER NOT NULL DEFAULT 0,
+    filter_editable INTEGER NOT NULL DEFAULT 0
+);
+CREATE TABLE feeds (
+    id INTEGER PRIMARY KEY,
+    name TEXT NOT NULL UNIQUE,
+    url TEXT NOT NULL,
+    enabled INTEGER NOT NULL DEFAULT 1,
+    last_success TEXT,
+    last_error TEXT
+);
+CREATE TABLE catalog_entries (
+    feed_id INTEGER NOT NULL REFERENCES feeds(id) ON DELETE CASCADE,
+    category TEXT NOT NULL,
+    service TEXT NOT NULL,
+    cidr TEXT NOT NULL,
+    PRIMARY KEY (feed_id, category, service, cidr)
+);
+CREATE TABLE user_networks (
+    user_id INTEGER NOT NULL REFERENCES users(id) ON DELETE CASCADE,
+    cidr TEXT NOT NULL UNIQUE,
+    PRIMARY KEY (user_id, cidr)
+);
+CREATE TABLE selected_categories (
+    user_id INTEGER NOT NULL REFERENCES users(id) ON DELETE CASCADE,
+    category TEXT NOT NULL,
+    PRIMARY KEY (user_id, category)
+);
+CREATE TABLE selected_services (
+    user_id INTEGER NOT NULL REFERENCES users(id) ON DELETE CASCADE,
+    category TEXT NOT NULL,
+    service TEXT NOT NULL,
+    PRIMARY KEY (user_id, category, service)
+);
+INSERT INTO users(id, name, peer_ip, peer_asn, filter_override_enabled)
+VALUES (7, 'legacy', '172.16.0.2', 65007, 1);
+INSERT INTO user_networks(user_id, cidr) VALUES (7, '192.168.20.0/24');
+`)
+	if err != nil {
+		t.Fatal(err)
+	}
+	if err := db.Close(); err != nil {
+		t.Fatal(err)
+	}
+
+	s, err := Open(path, config.Config{})
+	if err != nil {
+		t.Fatal(err)
+	}
+	defer s.Close()
+	user, err := s.User(context.Background(), 7)
+	if err != nil {
+		t.Fatal(err)
+	}
+	if user.FilterMode != FilterModeOverride || !user.FilterOverride {
+		t.Fatalf("legacy filter mode not preserved: %#v", user)
+	}
+}
