+++ internal/store/feeds_test.go
@@ -1,0 +1,89 @@
+package store
+
+import (
+	"context"
+	"database/sql"
+	"testing"
+
+	_ "modernc.org/sqlite"
+)
+
+func TestUpdateFeedURLClearsSnapshotAndDeleteCascades(t *testing.T) {
+	s := openTestStore(t)
+	ctx := context.Background()
+	if err := s.AddFeed(ctx, "custom", "https://example.test/old.json", true, 0); err != nil {
+		t.Fatal(err)
+	}
+	feeds, err := s.Feeds(ctx, false)
+	if err != nil {
+		t.Fatal(err)
+	}
+	feed := feeds[len(feeds)-1]
+	if _, err := s.DB.Exec(`
+UPDATE feeds SET last_success = 'now', last_error = 'old error' WHERE id = ?;
+INSERT INTO catalog_entries(feed_id, category, service, cidr)
+VALUES (?, 'Custom', 'Example', '8.8.8.0/24')`, feed.ID, feed.ID); err != nil {
+		t.Fatal(err)
+	}
+
+	feed.Name = "renamed"
+	feed.URL = "https://example.test/new.json"
+	if err := s.UpdateFeed(ctx, feed); err != nil {
+		t.Fatal(err)
+	}
+	feeds, err = s.Feeds(ctx, false)
+	if err != nil {
+		t.Fatal(err)
+	}
+	updated := feeds[len(feeds)-1]
+	if updated.Name != feed.Name || updated.URL != feed.URL ||
+		updated.LastSuccess != "" || updated.LastError != "" {
+		t.Fatalf("updated feed = %#v", updated)
+	}
+	var entries int
+	if err := s.DB.QueryRow("SELECT COUNT(*) FROM catalog_entries WHERE feed_id = ?", feed.ID).
+		Scan(&entries); err != nil {
+		t.Fatal(err)
+	}
+	if entries != 0 {
+		t.Fatalf("entries after URL change = %d, want 0", entries)
+	}
+
+	if _, err := s.DB.Exec(`INSERT INTO catalog_entries(feed_id, category, service, cidr)
+		VALUES (?, 'Custom', 'Example', '8.8.8.0/24')`, feed.ID); err != nil {
+		t.Fatal(err)
+	}
+	userID, err := s.AddUser(ctx, User{
+		Name: "selected", PeerIP: "172.16.0.2", PeerASN: 65001, Enabled: true,
+		Networks: []string{"192.168.20.0/24"},
+	})
+	if err != nil {
+		t.Fatal(err)
+	}
+	if err := s.Transaction(ctx, func(tx *sql.Tx) error {
+		return SetUserSelection(ctx, tx, userID, []string{"Custom"},
+			[]ServiceKey{{Category: "Custom", Service: "Example"}})
+	}); err != nil {
+		t.Fatal(err)
+	}
+	if err := s.DeleteFeed(ctx, feed.ID); err != nil {
+		t.Fatal(err)
+	}
+	if err := s.DB.QueryRow("SELECT COUNT(*) FROM catalog_entries WHERE feed_id = ?", feed.ID).
+		Scan(&entries); err != nil {
+		t.Fatal(err)
+	}
+	if entries != 0 {
+		t.Fatalf("entries after feed deletion = %d, want 0", entries)
+	}
+	for _, table := range []string{"selected_categories", "selected_services"} {
+		var selections int
+		if err := s.DB.QueryRow("SELECT COUNT(*) FROM "+table+" WHERE user_id = ?", userID).
+			Scan(&selections); err != nil {
+			t.Fatal(err)
+		}
+		if selections != 0 {
+			t.Fatalf("%s after feed deletion = %d, want 0", table, selections)
+		}
+	}
+}
